requirements.txt
catboost==1.2.8
fastapi==0.116.0
fastjsonschema==2.19.1
httpcore==1.0.4
httplib2==0.20.2
httpx==0.27.0
joblib==1.4.2
json5==0.9.17
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
jupyterlab_widgets==3.0.10
matplotlib==3.8.4
matplotlib-inline==0.1.6
numpy==1.26.4
pandas==2.2.1
pandocfilters==1.5.1
plotly==6.2.0
psycopg2-binary==2.9.10
pytest-lazy-fixture==0.6.3
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
sqlacodegen==3.0.0
SQLAlchemy==2.0.41
uvicorn==0.35.0
xgboost==3.0.2

